Adds `IntMap4` lookup structure

+++ TidalPy/utilities/lookups/intmap3.pyx
@@ -8,13 +8,13 @@ cdef class IntMap3:
     cdef void c_clear(self) noexcept nogil:
         self.intmap3_cinst.clear()
 
-    cdef void c_set(self, size_t l, size_t m, size_t p, double value) noexcept nogil:
+    cdef void c_set(self, int16_t l, int16_t m, int16_t p, double value) noexcept nogil:
         self.intmap3_cinst.set(l, m, p, value)
 
     cdef size_t c_size(self) noexcept nogil:
         return self.intmap3_cinst.size()
     
-    cdef cpp_bool c_get(self, double& result, size_t l, size_t m, size_t p) noexcept nogil:
+    cdef cpp_bool c_get(self, double& result, int16_t l, int16_t m, int16_t p) noexcept nogil:
         cdef cpp_bool found = False
         result = self.intmap3_cinst.get(found, l, m, p)
         return found
@@ -26,13 +26,13 @@ cdef class IntMap3:
     def clear(self):
         self.c_clear()
     
-    def set(self, size_t l, size_t m, size_t p, double value):
+    def set(self, int16_t l, int16_t m, int16_t p, double value):
         self.c_set(l, m, p, value)
 
     def size(self):
         return self.c_size()
     
-    def get(self, size_t l, size_t m, size_t p):
+    def get(self, int16_t l, int16_t m, int16_t p):
         cdef double result = 0.0
         cdef cpp_bool found = self.c_get(result, l, m, p)
         if not found:
@@ -43,17 +43,17 @@ cdef class IntMap3:
         if len(key) != 3:
             raise ValueError("Key must be a tuple of 3 integers (l, m, p)")
         # Unpack tuple
-        cdef size_t l = key[0]
-        cdef size_t m = key[1]
-        cdef size_t p = key[2]
+        cdef int16_t l = key[0]
+        cdef int16_t m = key[1]
+        cdef int16_t p = key[2]
         self.c_set(l, m, p, value)
 
     def __getitem__(self, tuple key):
         if len(key) != 3:
             raise ValueError("Key must be a tuple of 3 integers (l, m, p)")
-        cdef size_t l = key[0]
-        cdef size_t m = key[1]
-        cdef size_t p = key[2]
+        cdef int16_t l = key[0]
+        cdef int16_t m = key[1]
+        cdef int16_t p = key[2]
         cdef double result = 0.0
         cdef cpp_bool found = self.c_get(result, l, m, p)
         if not found:
